Move coerce_arrow into internals
To remove the public api dependency in utils.py
